+++ CHANGES.md
@@ -33,8 +33,8 @@ IronBee v0.8.0
 
 **Util**
 
-* Add external iterator support for hash.  See `ib_hash_iterator()`, `ib_hash_first()`, `ib_hash_next()`, `ib_hash_fetch()`, and `ib_hash_at_end()`.
-* `ib_hash_get()` and `ib_hash_get_ex()` now support NULL for the value argument to allow for membership tests.
+* Add external iterator support for hash.  See `ib_hash_iterator*`. `ib_hash_get()` and `ib_hash_get_ex()` now support NULL for the value argument to allow for membership tests.
+* Hash keys are now consistently `const char *` instead of a mix of `const char *` and `const void *`.
 * Hash now supports callback data for key hashing and equality.
 
 **IronBee++**